Migrate helper inside Evolution-Common and add License
I only keep the helpers that was used inside the demo-generator.
Put the license text inside every TypeScript file
Remove interviwerMonitoring that was not used.
