+++ .env.example
@@ -10,7 +10,6 @@ HOST = "http://localhost:8080"
 PG_DATABASE_DEVELOPMENT = "household_travel_survey_dev"
 PG_DATABASE_PRODUCTION = "household_travel_survey_prod"
 PG_DATABASE_TEST = "household_travel_survey_dev_test"
-# PG_DATABASE_Schema = "household_travel_survey"
 PG_CONNECTION_STRING_PREFIX = "postgres://postgres:<yourPassword>@localhost:5432/"
 
 # Express session
