+++ example/demo_generator/src/admin/app-admin.tsx
@@ -1,17 +1,20 @@
+/*
+ * Copyright 2025, Polytechnique Montreal and contributors
+ *
+ * This file is licensed under the MIT License.
+ * License text available at https://opensource.org/licenses/MIT
+ */
 import runClientApp from 'evolution-frontend/lib/apps/admin';
 import { setApplicationConfiguration } from 'chaire-lib-frontend/lib/config/application.config';
 import appConfig, { EvolutionApplicationConfiguration } from 'evolution-frontend/lib/config/application.config';
 
 import surveySections from '../survey/sections';
 import * as widgetsConfig from '../survey/widgetsConfigs';
-import projectHelpers from '../survey/helper';
-// import VisitedPlaceSection from '../survey/templates/VisitedPlacesSection';
 
 setApplicationConfiguration<EvolutionApplicationConfiguration>({
     sections: surveySections,
     widgets: widgetsConfig as any,
     allowedUrlFields: ['source', 'household.carNumber'],
-    projectHelpers: projectHelpers as any,
     templateMapping: { ...appConfig.templateMapping }
 });
