+++ example/demo_generator/src/server.ts
@@ -1,3 +1,9 @@
+/*
+ * Copyright 2025, Polytechnique Montreal and contributors
+ *
+ * This file is licensed under the MIT License.
+ * License text available at https://opensource.org/licenses/MIT
+ */
 import path from 'path';
 
 import setupServer from 'evolution-backend/lib/apps/participant/index';
@@ -7,14 +13,6 @@ import roleDefinitions from './survey/server/roleDefinition';
 import customFilterForValidation from './survey/server/serverConfiguration';
 import serverUpdateCallbacks from './survey/server/serverFieldUpdate';
 import serverValidations from './survey/server/serverValidations';
-
-// if (process.env.EV_VARIANT === undefined) {
-//     console.error('EV_VARIANT environment variable is not set. Please set it to the variant you want to run.');
-//     process.exit();
-// } else if (!surveyVariants.includes(process.env.EV_VARIANT)) {
-//     console.error('Invalid value for EV_VARIANT. Possible values are ', surveyVariants);
-//     process.exit();
-// }
 
 const configureServer = () => {
     // Default values for each field. Same as default config, but this is an example project, keep it here.
