+++ example/demo_generator/src/survey/server/serverConfiguration.ts
@@ -1,3 +1,9 @@
+/*
+ * Copyright 2025, Polytechnique Montreal and contributors
+ *
+ * This file is licensed under the MIT License.
+ * License text available at https://opensource.org/licenses/MIT
+ */
 import { defaultConfig } from 'evolution-backend/lib/config/projectConfig';
 import { InterviewListAttributes } from 'evolution-common/lib/services/questionnaire/types';
