+++ example/demo_generator/tests/common-tests-helpers.ts
@@ -1,3 +1,9 @@
+/*
+ * Copyright 2025, Polytechnique Montreal and contributors
+ *
+ * This file is licensed under the MIT License.
+ * License text available at https://opensource.org/licenses/MIT
+ */
 import * as testHelpers from 'evolution-frontend/tests/ui-testing/testHelpers';
 
 // Modify the CommonTestParameters type with survey parameters
